git_clone_corentin_01_07_2020/graphql-django-react/backend/requirements.txt
django
graphene_django
django-filter
django-cors-headers
graphene 
graphene-subscriptions
django-graphql-jwt
psycopg2
Pillow